requirements.txt

## The following requirements were added by pip freeze:
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-rest-registration==0.7.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7